MNT: move around the entrypoint a bit
